+++ CHANGELOG.md
@@ -132,6 +132,8 @@ DataFormats included:
 - Introduced 'Process' definitions (EIP pattern *Message Translator*)
 - Introduced action to clean headers (in Camel `<removeHeaders/>`)
 - New MAIL (SMTP) endpoint to send emails to mail server
+- New G-DRIVE endpoint to interact with Google Drive APIs
+- New G-SHEETS endpoint to interact with Google Sheets APIs
 - `<choice>` statement improvements:
   - Menu includes option for 1/2/3 conditions (+1 for `<otherwise>` branch) \
     If more needed, add them in the code and reload the visual editor.